- improved lazy evaluation of the references - removed duplicated & unused API
